Add configuration for protocol use in direct_link
We've been having problems with the production instance where the
connection between the reverse proxy and the web server does not
recognize the protocol (https front facing and http inside the network)
and the direct links to create plans were not working correctly.

At some point this should be addressed in deployment.