Set default funder when creating a new Plan
The current behavior of DMP Assistant does not require to select a funder when
creating a plan. Instead, a default funder is selected on the Plan controller
and that one is used to create the plan. This current implementation selects the
organization "Portage Network" in order to use its templates.

We will change this in the future to set this default organization on the user
interface.