Only submit invitation form if terms are accepted
This change results in the user being unable to submit the invitation form unless the accept terms checkbox is actually checked.
The checkbox handling here is a bit different from the checkbox handling within the create account section of the homepage (see the create method within RegistrationsController). We could do the same here, but it would require a bit more work. Specifically, we'd have to override the Devise::InvitationsController update method within InvitationsController. Also, a bit of extra handling would be required within the DeviseController require_no_authentication method that we are already overriding within InvitationsController.