Add * route to fix ActionController::RoutingError
This fix comes from the following link: https://rollbar.com/blog/ruby-on-rails-routingerror/

According to https://stackoverflow.com/a/37174557/17842948, "The ActionController::RoutingError is raised when Rails tries to match the request with a route. This happens before Rails even initializes a controller - thus your ApplicationController never has a chance to rescue the exception."

This fix prevents the ActionController::RoutingError by adding a catch-all/wildcard route to config/routes.rb.