DMP ASS'T PATCH Plan.organisationally_or_publicly_visible
This commit is intended to patch issue https://github.com/portagenetwork/roadmap/issues/732

This DMP Assistant fix was implemented in a way to minimise conflicts with the upstream, DMPRoadmap repo. This issue also exists in the upstream repo, so this commit will be readdressed after a fix is implemented there.