Fix "Create a new plan" features test
The "Create Plan" UI and logic For DMP Assistant diverges from that of DMP Roadmap. This commit adapts the tests within `spec/features/plans_spec.rb` to be compatible with DMP Assistant.

NOTE: I don't believe the `stub_openaire` call is required, so it has been removed for this test.