WIP SSO integration
This is a dirty work in progress commit that works under very specific
circumstances.

In this version, if I log in and there is a local account in the system
it will let me log in.

This needs to be further refined to handle all cases in a robust manner