Fix handling of `privileges_changed` assignment
When `privileges_changed == true`, the admin_privileges email is triggered. Prior to this commit, the triggered email was working properly with respect to removed privileges. However, in terms of added privileges, it was only being triggered when the API access privilege was added. This commit results in the email now being triggered after any privilege is added.