Add rake task for `4.1.1+portage-4.2.0` upgrade
Release `4.1.1+portage-4.2.0` includes a migration and it requires a specific new entry to the `identifier_schemes` table for openid_connect / CILogon to work properly. This rake task is intended to take care of all of that.

Note: The changes applied via this rake task have already been executed manually on the staging environment. However, performing the rake task on staging should still have no adverse effects.