Fix user creation on SSO signup
When deploying into production we noticed error messages where a user
account could not be found by email. When creating the account with
that email, the system crashed since email needs to be unique.

This change will fix the problem since it only updates the user instance
if it is a new record.