TEMP: customise devise.failure.unconfirmed values
This commit customises Devise's default value for the key `devise.failure.unconfirmed`. The value adds a link to the email confirmation page.
Note: The values for this key in both the `...en-CA.yml` and `...fr-CA.yml` files were provided solely by me and need to be approved/finalised by The Alliance