Update handling of default template in dropdown
This change resolves https://github.com/portagenetwork/roadmap/issues/1013
Before this change, `!customization.upgrade_customization?` was applied to all templates populating the dropdown, except the default template. As a result, if an organization had customized and published an older version of the default template, but not the newest version, that newest version would be blocked from being populated in the dropdown.