Add DB readiness check in `entrypoint.sh`
Adds a database readiness check in `entrypoint.sh` to ensure a DB connection before executing `bundle exec rake translation:sync`. This is necessary because our customized version of the translation gem requires DB access.

The script uses `bundle exec rails db:version` to verify the DB connection. It retries every 3 seconds, up to a maximum of 20 attempts (60 seconds total). If the database is still unreachable after this time, the script exits with an error (`exit 1`).

If the DB is reachable, the script runs `translation:sync`. If that task fails, the error is logged, but the app continues to start.

Finally, the Rails server is started with `bundle exec rails s`.