Remove `with_preferred_locale` from reset password view
Removing `<% LocaleService.with_preferred_locale(user) do %>` addresses an inconsistency with the delivered reset password instructions email. Specifically, the removed code was setting the email body to the user's preferred language, whereas the reset password instructions email subject was translating according to the app's current language.

Localising the email subject to the user's preferred locale would entail some extra Devise overriding. So for now, to keep things simple and consistent, we are also using the app's current locale to set the language of the email body.